refactor: OrderService로부터 주문 로직 분리
FDORA-145
